All code is now successfully compiled. Matching annotations are moved to commons module to be able to annotate the things we would like to search with @Matchable interface.

+++ commons/src/main/java/be/fooda/backend/commons/model/template/contact/request/FoodaAddressReq.java
@@ -5,9 +5,9 @@ import lombok.Data;
 
 @Data
 @Builder
-public class FoodaAddressReq { // sending post, put, delete, get request
-    private Integer addressId;
-    private Integer userId;
+public class FoodaAddressReq {
+    private Long addressId;
+    private Long userId;
     private String doorNo;
     private String number;
     private String street;
